Created a useCase hook
